+++ src/wiremind_kubernetes/__init__.py
@@ -1,6 +1,7 @@
 # noqa: F401
 from .kubernetes_helper import (  # noqa: F401
     KubernetesHelper,
+    NamespacedKubernetesHelper,
     KubernetesDeploymentManager,
 )
 from .utils import run_command  # noqa: F401